Navigation Update
Added Fullpage JS to Projects list page, updated navigation between projects....pending overhaul.
