+++ app/projects/[project]/page.tsx
@@ -44,7 +44,7 @@ export default async function Project({ params }: TypeProjectsFields) {
 
   const { items } = await Client.getEntries({
     content_type: 'projects',
-    'fields.slug': params!.project
+    'fields.projectId': params!.project
   })
 
   const projectData: TypeProjectsFields = items.map(
@@ -124,6 +124,44 @@ export default async function Project({ params }: TypeProjectsFields) {
                   </div>
                 </section>
               </main>
+
+              <footer>
+
+                <ol>
+
+                  <li>
+                    <Link
+                      // href={`/projects/${projectId}`}
+                      href={{
+                        pathname: `/projects/${project.projectId! + 1}`,
+                      }}
+                      as={`/projects/${project.projectId! + 1}`}
+                    >Next Project</Link>
+
+                  </li>
+
+                  <li>
+                    <Link
+                      // href={`/projects/${projectId}`}
+                      href={{
+                        pathname: `/projects/${project.projectId! + 1}`,
+                      }}
+                      as={`/projects/${project.projectId! - 1}`}
+                    >Previous Project</Link>
+
+                  </li>
+
+                  <li>
+                    <Link href={`/projects#${project.slug}`}>Back To Projects</Link>
+                  </li>
+
+                  <li>
+                    <Link href={'/'}>Back Home</Link>
+                  </li>
+                </ol>
+
+              </footer>
+
             </>
           </div>
         ))
